backend/requirements.txt
asgiref==3.11.1
astroid==4.0.3
attrs==25.4.0
autoflake==2.3.1
black==26.1.0
cfgv==3.5.0
click==8.3.1
dill==0.4.1
distlib==0.4.0
Django==5.2.11
django-cors-headers==4.9.0
django-filter==25.2
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.20.3
flake8==7.3.0
gprof2dot==2025.4.14
identify==2.6.16
inflection==0.5.1
isort==7.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pre_commit==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.11.0
pylint==4.0.4
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tomlkit==0.14.0
typing_extensions==4.15.0
uritemplate==4.2.0
virtualenv==20.36.1
Pillow>=10.0
